+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 
 ## v2.3.1-bcc (Jun XX, 2021)
 
+- fix bug with some stats in class filter
+- own class is now on top for class filter options
 - (bcc) add missing honor prices for horde
 - (bcc) add missing legendarys
 - (bcc) add missing world epics